worker/requirements.txt
celery==5.3.4
redis==5.0.1
pdfplumber==0.10.3
stix2==3.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
